- move interfaces related to VisualC to their own namespace
